Change initial value for reputation rating.
Change initial value for reputation rating from 0.5 to 0.0 for both
bidders. Justification: in the beginning (system warmup stage), both
bidders will experience perfect user success report list (with all
entries equal to 1); hence, when the window fills up, there is going to
be a jump in reputation rating from 0.5 to 0.0.